% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{checkFull}
\alias{checkFull}
\title{checkFull}
\usage{
checkFull(split.info, num.train, num.test, num.valid)
}
\arguments{
\item{split.info}{A list containing information such as data subscripts and classification subsets that need to be divided}

\item{num.train}{The sampling set of the training set}

\item{num.test}{The sampling set of the test set}

\item{num.valid}{The sampling set of the validation set}
}
\value{
A list
}
\description{
Built-in function: check whether sampling of any two subsets has been completed,
and all remaining data is allocated to the last subset
}
