% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{dataSplit}
\alias{dataSplit}
\title{dataSplit}
\usage{
dataSplit(data, control = list(), ...)
}
\arguments{
\item{data}{Raw data set entered by the user in matrix or data.frame format}

\item{control}{The list of parameters given by the user}

\item{...}{Redundant parameter list}
}
\value{
A list of three subsets
}
\description{
The main program entry of data splitting algorithms, provides the interface of each algorithm
Data format requirements:
Refer to the data sample provided by the package:The first column should be the index, the second through Nth columns should be the input vector,
and the last column should be the output vector
Refer to \code{\link{par.default}} for user-defined parameters.
}
\examples{
data("dataSplitAlgorithm_data_small")
result = dataSplit(dataSplitAlgorithm_data_small, control = list(sel.alg = "MDUPLEX",writeFile = FALSE))

data("dataSplitAlgorithm_data_middle")
result = dataSplit(dataSplitAlgorithm_data_middle, control = list(sel.alg = "SBSS.P",writeFile = FALSE))

data("dataSplitAlgorithm_data_large")
result = dataSplit(dataSplitAlgorithm_data_large, control = list(sel.alg = "SOMPLEX",writeFile = FALSE))

}
\references{
Chen, J., Zheng F., May R., Guo D., Gupta H., and Maier H. R.(2022).Improved data splitting methods for data-driven hydrological model development based on a large number of catchment samples, Journal of Hydrology, 613.

Zheng, F., Chen J., Maier H. R., and Gupta H.(2022). Achieving Robust and Transferable Performance for Conservation‐Based Models of Dynamical Physical Systems, Water Resources Research, 58(5).

Zheng, F., Chen, J., Ma, Y.,  Chen Q., Maier H. R., and Gupta H.(2023). A Robust Strategy to Account for Data Sampling Variability in the Development of Hydrological Models, Water Resources Research, 59(3).
}
\author{
Feifei Zheng \email{feifeizheng@zju.edu.cn}

Junyi Chen \email{jun1chen@zju.edu.cn}
}
