% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{DP.resample}
\alias{DP.resample}
\title{DP.resample}
\usage{
DP.resample(split.info, choice)
}
\arguments{
\item{split.info}{A list containing information such as data subscripts and classification subsets that need to be divided}

\item{choice}{The current sample set, a string, must have the same name as the subset in split.info}
}
\value{
A list to complete the current sampling
}
\description{
Built-in function: cyclic sampling of DUPLEX algorithm
}
