% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{getAUC}
\alias{getAUC}
\title{getAUC}
\usage{
getAUC(data1, data2)
}
\arguments{
\item{data1}{A matrix or data.frame after data splitting}

\item{data2}{A matrix or data.frame after data splitting}
}
\value{
a number, representing the average value of the auc
}
\description{
The similarity degree of distribution features between two data subsets are calculated and evaluated by auc index
}
