% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{getSampleNumofEachNeuron}
\alias{getSampleNumofEachNeuron}
\title{getSampleNumofEachNeuron}
\usage{
getSampleNumofEachNeuron(som.info, control)
}
\arguments{
\item{som.info}{A List which contains data points clustered by each neuron in the SOM network}

\item{control}{The list of parameters given by the user}
}
\value{
A list containing the sampling number of three subsets in the SOM grid
}
\description{
Built-in function: calculate the number of data segmentation for each neuron in the SOM grid
}
