% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{MDUPLEX}
\alias{MDUPLEX}
\title{Modified DUPLEX -- MDUPLEX algorithm}
\usage{
MDUPLEX(data, control)
}
\arguments{
\item{data}{Raw data set entered by the user in matrix or data.frame format}

\item{control}{The list of parameters given by the user}
}
\value{
A list of three subsets
}
\description{
DUPLEX have provided a mechanism for splitting the data in a deterministic manner,
a property that is attractive for practical applications. However,it can result in significant bias when used in the context of evaluating the performance of data-driven rainfall-runoff models where the size of the calibration subset is typically larger than that of the evaluation subset. Under such circumstances,
DUPLEX allocates a substantially larger number of normal (less extreme) data points to the larger (calibration) subset. This biases the calibration towards normal events and causes the model to have relatively poor performance on extreme events, resulting in pessimistic assessment of model evaluation performance
To overcome this problem, the authors of this package have developed an improved version of MDUPLEX algorithm, which can effectively solve the defects of traditional DUPLEX algorithm.
The details of the algorithm improvement are described in:
}
\references{
Chen, J., Zheng F., May R., Guo D., Gupta H., and Maier H. R.(2022), Improved data splitting methods for data-driven hydrological model development based on a large number of catchment samples, Journal of Hydrology, 613.

Zheng, F., Chen J., MaierH. R., and Gupta H.(2022), Achieving Robust and Transferable Performance for Conservation‐Based Models of Dynamical Physical Systems, Water Resources Research, 58(5).
}
