% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{par.default}
\alias{par.default}
\title{par.default}
\usage{
par.default()
}
\arguments{
\item{include.inp}{Whether the input vector is included when calculating Euclidean distance between samples, which defaults to TRUE}

\item{seed}{Seed number, which defaults to 1000}

\item{sel.alg}{Algorithm selection, includes SOMPLEX, MDUPLEX, DUPLEX, SBSS.P, TC, which defaults to "MDUPLEX"}

\item{prop.Tr}{The proportion of data allocated to the training set, which defaults to 0.6}

\item{prop.Ts}{The proportion of data allocated to the test set, which defaults to 0.2}

\item{Train}{The name of the training set when the output file is required, which defaults to "Train.txt"}

\item{Test}{The name of the test set when the output file is required, which defaults to "Test.txt"}

\item{Validation}{The name of the validation set when the output file is required, which defaults to "Valid.txt"}

\item{loc.calib}{If using the TC algorithm, you can also specify the location of the splitting range, which defaults to c(0,0.6)}

\item{writeFile}{Whether to output the partition result to a txt file, which defaults to TRUE}
}
\value{
None
}
\description{
The basic parameters defined by user
}
