% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{remainUnsample}
\alias{remainUnsample}
\title{remainUnsample}
\usage{
remainUnsample(X, Y)
}
\arguments{
\item{X}{A vector that needs to be sampled}

\item{Y}{A vector that have sampled}
}
\value{
A vector containing the remaining unsampled data
}
\description{
Built-in function: the auxiliary function of SS algorithm is used to obtain the remaining unsampled data
}
