% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{SBSS.P}
\alias{SBSS.P}
\title{SBSS-P algorithm}
\usage{
SBSS.P(data, control)
}
\arguments{
\item{data}{Raw data set entered by the user in matrix or data.frame format}

\item{control}{The list of parameters given by the user}
}
\value{
A list of three subsets
}
\description{
The self-organizing map based stratified sampling(SBSS) approach is a two-step data splitting method,
which has been implemented previously by (Bowden et al., 2002), (Kingston, 2006), and (May et al., 2010).
In the first step, multivariate stratified random sampling is performed to partition the data into K strata, where clustering is performed using a self-organizing map (SOM,Kohonen, 1998).
In the second step, uniform random intracluster sampling is applied to generate the data split.
}
\references{
May, R. J., Maier H. R., and Dandy G. C.(2010), Data splitting for artificial neural networks using SOM-based stratified sampling, Neural Netw, 23(2), 283-294.
}
