% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{selectData}
\alias{selectData}
\title{selectData}
\usage{
selectData(data, control)
}
\arguments{
\item{data}{Raw data set entered by the user in matrix or data.frame format}

\item{control}{The list of parameters given by the user}
}
\value{
Remove the subscript vector in the first column and return a matrix
}
\description{
Built-in function: select data for splitting and decide whether to include input vectors based on
user-set parameter \code{include.inp}
}
