% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{somCluster}
\alias{somCluster}
\title{somCluster}
\usage{
somCluster(data)
}
\arguments{
\item{data}{Given data set in matrix or data.frame format}
}
\value{
A List which contains data points clustered by each neuron in the SOM network
}
\description{
Built-in function: SOM cluster by using \code{[kohonen]{som}}
}
