% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{SOMPLEX}
\alias{SOMPLEX}
\title{SOMPLEX algorithm}
\usage{
SOMPLEX(data, control)
}
\arguments{
\item{data}{Raw data set entered by the user in matrix or data.frame format}

\item{control}{The list of parameters given by the user}
}
\value{
A list of three subsets
}
\description{
SOMPLEX is a multi-stage algorithm that combines SOM-based clustering with DUPLEX sampling of map units.
SOM clustering is used to identify distinct regions of the data space that can include both typical classes and examples of unique input–output cases.
However, overlooked by many data splitting applications is that the clusters may cover varying proportions of the database, data can be non-uniformly distributed within each partition,
and data in some partitions can be more widely spread than in others (May et al., 2010), due to which random sampling from within each cluster can be less than ideal.
Conversely, given any distribution of data, DUPLEX can generate a sample that uniformly covers the full range of the data (Snee, 1977).
By applying DUPLEX to each SOM-based partition, representative samples can be generated from each partition regardless of the distribution within the partition.
}
\references{
Chen, J., Zheng F., May R., Guo D., Gupta H., and Maier H. R.(2022), Improved data splitting methods for data-driven hydrological model development based on a large number of catchment samples, Journal of Hydrology, 613.
}
