% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{SS}
\alias{SS}
\title{SS algorithm}
\usage{
SS(data, control)
}
\arguments{
\item{data}{Raw data set entered by the user in matrix or data.frame format}

\item{control}{The list of parameters given by the user}
}
\value{
A list of three subsets
}
\description{
The Systematic stratified(SS) data splitting method is a semideterministic method,
in which every kth sample from a random starting point is selected to form the training,
testing, and validation data sets. In implementing systematic sampling in this study,
the data are first ordered along the output variable dimension in increasing order.
Then, the sampling interval is determined based on the training and testing data proportions specified by the user.
Thereafter, a starting point is randomly selected and training samples are drawn first, followed by the testing samples.
Finally, unsampled data are allocated to the validation set.
}
\references{
Baxter, C. W., Stanley S. J., Zhang Q., and Smith D. W.(2000), Developing artificial neural network process models: A guide for drinking water utilities,

paper presented at the 6th Environmental Engineering Society Specialty Conference of the CSCE, pp. 376–383, Canadian Society for Civil Engineering (CSCE),

London, Ontario, Canada.
}
