% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{SSsample}
\alias{SSsample}
\title{SSsample}
\usage{
SSsample(index, prop)
}
\arguments{
\item{index}{A vector containing the index of the data}

\item{prop}{A number determining the sampling step}
}
\value{
A vector with sampled data index
}
\description{
Built-in function: core code of SS algorithm
}
