% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{standardise}
\alias{standardise}
\title{standardise}
\usage{
standardise(data)
}
\arguments{
\item{data}{Enter data in matrix or data.frame format}
}
\value{
A matrix with the internal data normalized by column
}
\description{
Build-in function: using for standardizing the raw data
}
