% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplitAlgorithm.R
\name{TC}
\alias{TC}
\title{TC algorithm}
\usage{
TC(data, control)
}
\arguments{
\item{data}{Raw data set entered by the user in matrix or data.frame format}

\item{control}{The list of parameters given by the user}
}
\value{
A list containing subsets of calibration and evaluation data
}
\description{
Traditional data splitting algorithm: simply divide the raw data into two parts for calibration and evaluation
}
